requirements.txt
pytest>=8.0.0
pytest-asyncio>=1.0.0
aiohttp>=3.9.0
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0